requirements.txt
pytest==7.3.1
requests==2.30.0
python-dotenv==1.0.0